move makepath to data_manager
